+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/data_sources/remote/count_remote_data_source.dart
@@ -8,15 +8,14 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
 
-import '../../../../base/models/count.dart';
-import 'count_data_source.dart';
+import '../../models/count.dart';
 
 part 'count_remote_data_source.g.dart';
 
 /// Used as a contractor for remote data source.
 /// To make it work, should provide a real API and rerun build_runner
-@RestApi(baseUrl: 'https://api.primeholding.com')
-abstract class CountRemoteDataSource implements CountDataSource{
+@RestApi(baseUrl: 'http://0.0.0.0:8080/api')
+abstract class CountRemoteDataSource{
   factory CountRemoteDataSource(Dio dio,
       {String baseUrl}) = _CountRemoteDataSource;
 
@@ -24,11 +23,11 @@ abstract class CountRemoteDataSource implements CountDataSource{
   @override
   Future<Count> getCurrent();
 
-  @PUT('/count/increment')
+  @POST('/count/increment')
   @override
   Future<Count> increment();
 
-  @PUT('/count/decrement')
+  @POST('/count/decrement')
   @override
   Future<Count> decrement();
 }
