+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_counter/di/counter_dependencies.dart
@@ -10,9 +10,7 @@ import 'package:flutter_rx_bloc/flutter_rx_bloc.dart';
 import 'package:provider/provider.dart';
 import 'package:provider/single_child_widget.dart';
 
-import '../../base/data_sources/domain/counter/count_data_source.dart';
-import '../../base/data_sources/domain/counter/count_local_data_source.dart';
-import '../../base/data_sources/domain/counter/count_remote_data_source.dart';
+import '../../base/data_sources/remote/count_remote_data_source.dart';
 import '../../base/repositories/counter_repository.dart';
 import '../blocs/counter_bloc.dart';
 
@@ -40,9 +38,8 @@ class CounterDependencies {
   /// lib\base\data_sources\domain\counter\count_remote_data_source.dart
   /// and data models in lib\base\models and rerun build_runner.
   List<Provider> get _dataSources => [
-    Provider<CountDataSource>(
-      // create: (context) => CountRemoteDataSource(context.read()),
-      create: (context) => CountLocalDataSource(),
+    Provider<CountRemoteDataSource>(
+      create: (context) => CountRemoteDataSource(context.read()),
     ),
       ];
