+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/test/feature_counter/blocs/counter_bloc_test.mocks.dart
@@ -1,13 +1,14 @@
-// Mocks generated by Mockito 5.0.9 from annotations
+// Mocks generated by Mockito 5.0.10 from annotations
 // in test_app/test/feature_counter/blocs/counter_bloc_test.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i5;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:{{project_name}}/base/data_sources/domain/counter/count_data_source.dart' as _i2;
-import 'package:{{project_name}}/base/models/count.dart' as _i3;
-import 'package:{{project_name}}/base/repositories/counter_repository.dart' as _i4;
+import 'package:test_app/base/data_sources/remote/count_remote_data_source.dart'
+as _i2;
+import 'package:test_app/base/models/count.dart' as _i3;
+import 'package:test_app/base/repositories/counter_repository.dart' as _i4;
 
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: comment_references
@@ -15,7 +16,8 @@ import 'package:{{project_name}}/base/repositories/counter_repository.dart' as _
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 
-class _FakeCountDataSource extends _i1.Fake implements _i2.CountDataSource {}
+class _FakeCountRemoteDataSource extends _i1.Fake
+    implements _i2.CountRemoteDataSource {}
 
 class _FakeCount extends _i1.Fake implements _i3.Count {}
 
@@ -28,9 +30,9 @@ class MockCounterRepository extends _i1.Mock implements _i4.CounterRepository {
   }
 
   @override
-  _i2.CountDataSource get countDataSource =>
-      (super.noSuchMethod(Invocation.getter(#countDataSource),
-          returnValue: _FakeCountDataSource()) as _i2.CountDataSource);
+  _i2.CountRemoteDataSource get countDataSource => (super.noSuchMethod(
+      Invocation.getter(#countDataSource),
+      returnValue: _FakeCountRemoteDataSource()) as _i2.CountRemoteDataSource);
   @override
   _i5.Future<_i3.Count> getCurrent() =>
       (super.noSuchMethod(Invocation.method(#getCurrent, []),
